Adding team_shot_put exercise.

+++ exercises/pthreads/grandma_lottery_1/src/grandma_lottery_1.c
@@ -26,8 +26,8 @@ int main(void) {
                                         /*function arguments*/ 2);
     if (grandchildren1 == EXIT_SUCCESS && grandchildren2 == EXIT_SUCCESS) {
         printf("What numbers did you get?\n");
-        long int* retval1;
-        long int* retval2;
+        long long int* retval1;
+        long long int* retval2;
         pthread_join(thread, &retval1);
         pthread_join(thread2, &retval2);
 
@@ -42,9 +42,9 @@ int main(void) {
 // end procedure
 
 void* grandchildren(void* data) {
-    long int number = (long int*)data;
+    long long int number = (long long int*)data;
     // Utilizing rand() as it doesn't need a seed and rand_r has been deprecated
-    long int* randint = (long int)rand() % 100;
+    long long int* randint = (long long int)rand() % 100;
     printf("I got the number: %li (%li)\n", randint, number);
     return randint;
 }
